Use int8 to limit indent width

+++ src/json.h
@@ -913,9 +913,8 @@ namespace json {
 struct StringifyOptions
 {
     // If >= 0, pretty-print the JSON.
-    // Default is < 0, that is the JSON is rendered as the shortest string
-    // possible.
-    int indent_width = -1;
+    // Default is < 0, that is the JSON is rendered as the shortest string possible.
+    int8_t indent_width = -1;
 
     // If true, check for invalid UTF-8 sequences in strings.
     // Default is true.
